Moved did_already_insert_ad to a new class
